Stub out the detail view with a PieceMeta $resource
Use @id to extract the 'id' value from the data object, instead of
passing it in as a separate parameter [1].  We don't have an :id
marker in the search URL, so it doesn't matter there if we extract an
extra parameter.

There's also some fun with $location.path to push folks back to the
queue after a successful save.  One less click for folks grinding
through the queue.

[1]: https://docs.angularjs.org/api/ngResource/service/$resource
[2]: https://docs.angularjs.org/api/ng/service/$location