QueueCtrl: Add paging view
It takes a while to load all eight thousand PieceMetas, and nobody is
going to read through that whole list at once.  Use Beehive's new
'from', 'size', and 'count' query parameters to setup a paging view,
so we only need to fetch a handful of PieceMetas at a time.

I use parseInt [1] to convert the parameters to integers, otherwise
'from + size' can be string addition (and "4" + 10 = "410" is probably
not what you want ;).

I also use Math.max [2] to avoid starting at a negative position,
Math.min [3] to avoid "from 0 to 10 of 4", and <ol start="..."> to
adjust the list numbering [4].

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/parseInt
     From ECMAScript 1st Edition and supported by all major browsers.
[2]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/max
     From ECMAScript 1st Edition and supported by all major browsers.
[3]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/min
     From ECMAScript 1st Edition and supported by all major browsers.
[4]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/ol#Using_the_start_attribute
     Part of HTML 4.01 and HTML 5.