QueueCtrl: Add claim buttons (and display queue as a table)
Now you can click the claim button to claim pieces in the queue view.
I use toISOString [1] to get current date strings, and the 'update'
command [2] to post the new information to Elasticsearch.

As of this commit,

  .then(function(result) {
    $scope.search();
  }

is getting stale search results.  I fix this in a later commit which
adds 'refresh' arguments to the updates.  Another issue with this UI
is that that claimed items are shunted to the back of our
arbitrarily-ordered search results (perhaps they are ordered by
edit-time?).

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toISOString
     Supported by major browsers and IE 9+.  Specified in ECMAScript
     5.1.
[2]: http://www.elasticsearch.org/guide/en/elasticsearch/client/javascript-api/current/api-reference.html#api-update