QueueCtrl: Redirect from the claim button to the detail view
Avoid the "claimed items shifted to end of search results" issue
mentioned in e5e8abc4 (QueueCtrl: Add claim buttons (and display queue
as a table, 2014-08-19).  A better alternative would be to add a
'priority' field to the Elasticsearch entries and then use that field
to order the search results.  Food for a later commit.