Add elasticsearch as a dependency
Three new dependencies: elasticsearch, es and es_client.

es_client is a factory similiar to what is found in the docs for
angular[1].

es is a wrapper around es_client to provide some convenience
functions with the ability to default _index and _type, as opposed
to having to pass them each time.  For now, making use of callbacks.
May want to move toward promises later, if needed.

The idea is to load the grids and individual items for review from
elasticsearch.  Changes would be persisted back to elasticsearch
until the "review" is complete for a given item.  At that point,
the changes would be persisted to a REST-ish endpoint as described
in config.

[1] http://www.elasticsearch.org/guide/en/elasticsearch/client/javascript-api/current/browser-builds.html#_angular_build